#136 chore: refatoracao mailer e ajuste testes
