+++ src/factories/useCases/recoverUserPassword.ts
@@ -1,9 +1,11 @@
 import UserRepository from '../../repository/userRepository'
 import { RecoverUserPasswordUseCase } from '../../useCase/recoverUserPassword/recoverUserPasswordUseCase'
 import { BcryptAdapter } from '../../adapters/bcryptAdapter'
+import { MailerAdapter } from '../../adapters/mailerAdapter'
 
 export const makeRecoverUserPassword = () => {
   const userRepository = new UserRepository()
   const encryptor = new BcryptAdapter()
-  return new RecoverUserPasswordUseCase(userRepository, encryptor)
+  const mailer = new MailerAdapter()
+  return new RecoverUserPasswordUseCase(userRepository, encryptor, mailer)
 }
