+++ src/useCase/authenticationUser/authenticationUserUseCase.spec.ts
@@ -26,7 +26,8 @@ const mockedUser: User = {
   role: Role.ADMIN,
   password: datatype.string(),
   createdAt: new Date(),
-  updatedAt: new Date()
+  updatedAt: new Date(),
+  temporaryPassword: false
 }
 
 const authenticateUserUseCase = new AuthenticateUserUseCase(
