+++ src/useCase/recoverUserPassword/recoverUserPasswordUseCase.ts
@@ -2,7 +2,7 @@ import { UseCase, UseCaseReponse } from '../protocols/useCase'
 import { Repository } from '../../repository/protocol/repository'
 import { Encryptor } from '../../services/encryptor'
 import { User } from '../../db/entities/user'
-import { sendEmail } from '../../services/mailer'
+import { MailService } from '../../services/mailer'
 import crypto from 'crypto'
 
 export interface RecoverUserPasswordData {
@@ -20,18 +20,17 @@ export class RecoverUserPasswordUseCase
   implements
     UseCase<{
       message: string
-      user: User
     }>
 {
   constructor(
     private readonly userRepository: Repository,
-    private readonly encryptor: Encryptor
+    private readonly encryptor: Encryptor,
+    private readonly mailer: MailService
   ) {}
 
   async execute(recoverPassword: RecoverUserPasswordData): Promise<
     UseCaseReponse<{
       message: string
-      user: User
     }>
   > {
     const response: User | undefined = await this.userRepository.findOneByEmail(
@@ -44,19 +43,22 @@ export class RecoverUserPasswordUseCase
       }
     const temporaryPassword = crypto.randomBytes(4).toString('hex')
     const hashedPassword = this.encryptor.encrypt(temporaryPassword)
-    if (await sendEmail(recoverPassword.email, temporaryPassword)) {
+    if (
+      await this.mailer.sendRecoverPasswordEmail(
+        recoverPassword.email,
+        temporaryPassword
+      )
+    ) {
       await this.userRepository.updateOne({
         userId: response.id,
         password: hashedPassword,
         temporaryPassword: true
       })
     }
-    this.userRepository.findOneByEmail(recoverPassword.email)
     return {
       isSuccess: true,
       data: {
-        message: 'Email enviado com sucesso',
-        user: response
+        message: 'Email enviado com sucesso'
       }
     }
   }
